Move UnusedPortFinder into TestRunner.
