+++ pwiz_tools/Skyline/TestRunner/TestRunner.csproj
@@ -113,6 +113,7 @@
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Statistics.cs" />
+    <Compile Include="UnusedPortFinder.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
